+++ uniffi_bindgen/src/lib.rs
@@ -97,10 +97,10 @@ const BINDGEN_VERSION: &str = env!("CARGO_PKG_VERSION");
 use anyhow::{anyhow, bail, Context, Result};
 use camino::{Utf8Path, Utf8PathBuf};
 use fs_err::{self as fs, File};
-use serde::{Deserialize, Serialize};
+use serde::{de::DeserializeOwned, Deserialize, Serialize};
 use std::io::prelude::*;
 use std::io::ErrorKind;
-use std::{collections::HashMap, env, process::Command, str::FromStr};
+use std::{env, process::Command, str::FromStr};
 
 pub mod backend;
 pub mod bindings;
@@ -111,70 +111,40 @@ pub mod scaffolding;
 pub use interface::ComponentInterface;
 use scaffolding::RustScaffolding;
 
-/// A trait representing a Binding Generator Configuration
+/// Trait for bindings configuration.  Each bindings language defines one of these.
 ///
-/// External crates that implement binding generators need to implement this trait and set it as
-/// the `BindingGenerator.config` associated type.  `generate_external_bindings()` then uses it to
-/// generate the config that's passed to `BindingGenerator.write_bindings()`
-pub trait BindingGeneratorConfig: for<'de> Deserialize<'de> {
-    /// Get the entry for this config from the `bindings` table.
-    fn get_entry_from_bindings_table(bindings: &toml::Value) -> Option<toml::Value>;
+/// BindingsConfigs are initially loaded from `uniffi.toml` file.  Then the trait methods are used
+/// to fill in missing values.
+pub trait BindingsConfig: DeserializeOwned {
+    /// key in the `bindings` table from `uniffi.toml` for this configuration
+    const TOML_KEY: &'static str;
 
-    /// Get default config values from the `ComponentInterface`
-    ///
-    /// These will replace missing entries in the bindings-specific config
-    fn get_config_defaults(ci: &ComponentInterface) -> Vec<(String, toml::Value)>;
+    /// Update missing values using the `ComponentInterface`
+    #[allow(unused)]
+    fn update_from_ci(&mut self, ci: &ComponentInterface) {}
 }
 
-fn load_bindings_config<BC: BindingGeneratorConfig>(
+fn load_bindings_config<BC: BindingsConfig>(
     ci: &ComponentInterface,
     crate_root: &Utf8Path,
     config_file_override: Option<&Utf8Path>,
 ) -> Result<BC> {
     // Load the config from the TOML value, falling back to an empty map if it doesn't exist
-    let mut config_map: toml::value::Table =
-        match load_bindings_config_toml::<BC>(crate_root, config_file_override)? {
-            Some(value) => value
-                .try_into()
-                .context("Bindings config must be a TOML table")?,
-            None => toml::map::Map::new(),
-        };
+    let toml_config = load_bindings_config_toml(crate_root, config_file_override)?
+        .and_then(|mut v| v.as_table_mut().and_then(|t| t.remove(BC::TOML_KEY)))
+        .unwrap_or_else(|| toml::Value::from(toml::value::Table::default()));
 
-    // Update it with the defaults from the component interface
-    for (key, value) in BC::get_config_defaults(ci) {
-        config_map.entry(key).or_insert(value);
-    }
-
-    // Leverage serde to convert toml::Value into the config type
-    toml::Value::from(config_map)
-        .try_into()
-        .context("Generating bindings config from toml::Value")
+    let mut config: BC = toml_config.try_into()?;
+    config.update_from_ci(ci);
+    Ok(config)
 }
 
 /// Binding generator config with no members
-#[derive(Clone, Debug, Hash, PartialEq, PartialOrd, Ord, Eq)]
-pub struct EmptyBindingGeneratorConfig;
+#[derive(Clone, Debug, Deserialize, Hash, PartialEq, PartialOrd, Ord, Eq)]
+pub struct EmptyBindingsConfig;
 
-impl BindingGeneratorConfig for EmptyBindingGeneratorConfig {
-    fn get_entry_from_bindings_table(_bindings: &toml::Value) -> Option<toml::Value> {
-        None
-    }
-
-    fn get_config_defaults(_ci: &ComponentInterface) -> Vec<(String, toml::Value)> {
-        Vec::new()
-    }
-}
-
-// EmptyBindingGeneratorConfig is a unit struct, so the `derive(Deserialize)` implementation
-// expects a null value rather than the empty map that we pass it.  So we need to implement
-// `Deserialize` ourselves.
-impl<'de> Deserialize<'de> for EmptyBindingGeneratorConfig {
-    fn deserialize<D>(_deserializer: D) -> Result<Self, D::Error>
-    where
-        D: serde::Deserializer<'de>,
-    {
-        Ok(EmptyBindingGeneratorConfig)
-    }
+impl BindingsConfig for EmptyBindingsConfig {
+    const TOML_KEY: &'static str = "";
 }
 
 // Load the binding-specific config
@@ -184,7 +154,7 @@ impl<'de> Deserialize<'de> for EmptyBindingGeneratorConfig {
 //
 // If there is an error parsing the file then Err will be returned. If the file is missing or the
 // entry for the bindings is missing, then Ok(None) will be returned.
-fn load_bindings_config_toml<BC: BindingGeneratorConfig>(
+fn load_bindings_config_toml(
     crate_root: &Utf8Path,
     config_file_override: Option<&Utf8Path>,
 ) -> Result<Option<toml::Value>> {
@@ -199,12 +169,12 @@ fn load_bindings_config_toml<BC: BindingGeneratorConfig>(
 
     let contents = fs::read_to_string(&config_path)
         .with_context(|| format!("Failed to read config file from {config_path}"))?;
-    let full_config = toml::Value::from_str(&contents)
+    let mut full_config = toml::Value::from_str(&contents)
         .with_context(|| format!("Failed to parse config file {config_path}"))?;
 
     Ok(full_config
-        .get("bindings")
-        .and_then(BC::get_entry_from_bindings_table))
+        .as_table_mut()
+        .and_then(|t| t.remove("bindings")))
 }
 
 /// A trait representing a UniFFI Binding Generator
@@ -212,9 +182,8 @@ fn load_bindings_config_toml<BC: BindingGeneratorConfig>(
 /// External crates that implement binding generators, should implement this type
 /// and call the [`generate_external_bindings`] using a type that implements this trait.
 pub trait BindingGenerator: Sized {
-    /// Associated type representing a the bindings-specifig configuration parsed from the
-    /// uniffi.toml
-    type Config: BindingGeneratorConfig;
+    /// Handles configuring the bindings
+    type Config: BindingsConfig;
 
     /// Writes the bindings to the output directory
     ///
@@ -264,16 +233,11 @@ pub fn generate_external_bindings(
 // such as the `extern "C"` function definitions and record data types.
 pub fn generate_component_scaffolding(
     udl_file: &Utf8Path,
-    config_file_override: Option<&Utf8Path>,
+    _config_file_override: Option<&Utf8Path>,
     out_dir_override: Option<&Utf8Path>,
     format_code: bool,
 ) -> Result<()> {
     let component = parse_udl(udl_file)?;
-    let _config = get_config(
-        &component,
-        guess_crate_root(udl_file)?,
-        config_file_override,
-    );
     let file_stem = udl_file.file_stem().context("not a file")?;
     let filename = format!("{file_stem}.uniffi.rs");
     let out_path = get_out_dir(udl_file, out_dir_override)?.join(filename);
@@ -301,7 +265,8 @@ pub fn generate_bindings(
     }
     let crate_root = &guess_crate_root(udl_file).context("Failed to guess crate root")?;
 
-    let config = get_config(&component, crate_root, config_file_override)?;
+    let mut config = Config::load_initial(crate_root, config_file_override)?;
+    config.update_from_ci(&component);
     let out_dir = get_out_dir(udl_file, out_dir_override)?;
     for language in target_languages {
         bindings::write_bindings(
@@ -341,29 +306,6 @@ pub fn guess_crate_root(udl_file: &Utf8Path) -> Result<&Utf8Path> {
         bail!("UDL file does not appear to be inside a crate")
     }
     Ok(path_guess)
-}
-
-fn get_config(
-    component: &ComponentInterface,
-    crate_root: &Utf8Path,
-    config_file_override: Option<&Utf8Path>,
-) -> Result<Config> {
-    let default_config: Config = component.into();
-
-    let config_file = match config_file_override {
-        Some(cfg) => Some(cfg.to_owned()),
-        None => crate_root.join("uniffi.toml").canonicalize_utf8().ok(),
-    };
-
-    match config_file {
-        Some(path) => {
-            let contents = fs::read_to_string(&path).context("Failed to read config file")?;
-            let loaded_config: Config = toml::de::from_str(&contents)
-                .with_context(|| format!("Failed to generate config from file {path}"))?;
-            Ok(loaded_config.merge_with(&default_config))
-        }
-        None => Ok(default_config),
-    }
 }
 
 fn get_out_dir(udl_file: &Utf8Path, out_dir_override: Option<&Utf8Path>) -> Result<Utf8PathBuf> {
@@ -406,44 +348,30 @@ struct Config {
     bindings: bindings::Config,
 }
 
-impl From<&ComponentInterface> for Config {
-    fn from(ci: &ComponentInterface) -> Self {
-        Config {
-            bindings: ci.into(),
-        }
+impl Config {
+    fn load_initial(
+        crate_root: &Utf8Path,
+        config_file_override: Option<&Utf8Path>,
+    ) -> Result<Self> {
+        let path = match config_file_override {
+            Some(cfg) => Some(cfg.to_owned()),
+            None => crate_root.join("uniffi.toml").canonicalize_utf8().ok(),
+        };
+        let toml_config = match path {
+            Some(path) => {
+                let contents = fs::read_to_string(path).context("Failed to read config file")?;
+                toml::de::from_str(&contents)?
+            }
+            None => toml::Value::from(toml::value::Table::default()),
+        };
+        Ok(toml_config.try_into()?)
     }
-}
 
-pub trait MergeWith {
-    fn merge_with(&self, other: &Self) -> Self;
-}
-
-impl MergeWith for Config {
-    fn merge_with(&self, other: &Self) -> Self {
-        Config {
-            bindings: self.bindings.merge_with(&other.bindings),
-        }
-    }
-}
-
-impl<T: Clone> MergeWith for Option<T> {
-    fn merge_with(&self, other: &Self) -> Self {
-        match (self, other) {
-            (Some(_), _) => self.clone(),
-            (None, Some(_)) => other.clone(),
-            (None, None) => None,
-        }
-    }
-}
-
-impl<V: Clone> MergeWith for HashMap<String, V> {
-    fn merge_with(&self, other: &Self) -> Self {
-        let mut merged = HashMap::new();
-        // Iterate through other first so our keys override theirs
-        for (key, value) in other.iter().chain(self) {
-            merged.insert(key.clone(), value.clone());
-        }
-        merged
+    fn update_from_ci(&mut self, ci: &ComponentInterface) {
+        self.bindings.kotlin.update_from_ci(ci);
+        self.bindings.swift.update_from_ci(ci);
+        self.bindings.python.update_from_ci(ci);
+        self.bindings.ruby.update_from_ci(ci);
     }
 }